Move Gives/Gains Experience and Bounty to Mods.Common
