+++ OpenRA.Mods.Common/Effects/Rank.cs
@@ -11,8 +11,9 @@
 using System.Collections.Generic;
 using OpenRA.Effects;
 using OpenRA.Graphics;
+using OpenRA.Mods.Common.Traits;
 
-namespace OpenRA.Mods.RA.Effects
+namespace OpenRA.Mods.Common.Effects
 {
 	class Rank : IEffect
 	{
