+++ OpenRA.Mods.RA/Traits/Crates/LevelUpCrateAction.cs
@@ -9,6 +9,7 @@
 #endregion
 
 using System.Linq;
+using OpenRA.Mods.Common.Traits;
 
 namespace OpenRA.Mods.RA.Traits
 {
